requirements.txt
Django==3.2.16
djangorestframework==3.14.0
requests==2.32.3
dotenv==0.9.9